Add type definitions
