+++ CopilotKit/packages/react-core/src/types/index.ts
@@ -1,3 +1,4 @@
 export type { AnnotatedFunctionArgument } from "./annotated-function";
 export type { AnnotatedFunction } from "./annotated-function";
 export type { DocumentPointer } from "./document-pointer";
+export type { Role, Message, Function, FunctionCall, FunctionCallHandler } from "./base";
